Initiating the Git repository.
wrote a function for API connect

trying to create a list of the video ids

Created the function to get a particular channels ids

trying to create a function to get video details

created the function to get required specific details of a particular video in a achannel.
Also, to note: 'youtube.quota'exceeded.
trying to fix this by limiting the number of api.execute()